+++ src/plone/restapi/services/auth/login.py
@@ -4,9 +4,7 @@ from Acquisition import aq_parent
 from plone.restapi.deserializer import json_body
 from plone.restapi.services import Service
 from Products.CMFCore.utils import getToolByName
-from Products.PluggableAuthService.interfaces.plugins import (
-    IAuthenticationPlugin,
-)
+from Products.PluggableAuthService.interfaces.plugins import IAuthenticationPlugin
 from zope.interface import alsoProvides
 
 import plone.protect.interfaces