+++ CHANGES.rst
@@ -8,6 +8,26 @@ Changelog
 
 .. towncrier release notes start
 
+5.0.2 (2019-11-06)
+------------------
+
+Bug fixes:
+
+
+- Fix filtering vocabs and sources by title with non-ASCII characters.
+  [lgraf] (#825)
+
+
+5.0.1 (2019-11-05)
+------------------
+
+Bug fixes:
+
+
+- Fix serialization of vocabulary items for fields that need hashable items (e.g. sets).
+  [buchi] (#788)
+
+
 5.0.0 (2019-10-31)
 ------------------